General cleanup and file move.
