+++ src/main/java/org/smack/util/converters/StringConverterExtension.java
@@ -1,0 +1,12 @@
+package org.smack.util.converters;
+
+/**
+ *
+ *
+ * @version $Revision$
+ * @author Michael Binz
+ */
+public abstract class StringConverterExtension
+{
+    public abstract void extendTypeMap( StringConverter registry );
+}
