--- src/main/java/org/smack/util/resource/ResourceConverterExtension.java
@@ -1,12 +1,0 @@
-package org.smack.util.resource;
-
-/**
- *
- *
- * @version $Revision$
- * @author Michael Binz
- */
-public abstract class ResourceConverterExtension
-{
-    public abstract void extendTypeMap( ResourceConverterRegistry registry );
-}
